ok on to something here
Found a populating union solution list error in fullsolution word.  The
idea is i do check for unions and populate the list correctly but only
in one direction.  I need to compair all pieces with all pieces in a
given pair list to find the complete union list that finds the 25 pieces
to solve puzzle.  I currently seem to be only finding the first x amount
that fill the union list untill the pair list is done but because i did
not find the solution i go on to the next pair list and do not check all
the combination pieces in that pair list for unions ... so this is why
the current fullsolution does not find an answer!